feat: add text rendering, correct group hierarchy
